+++ src/py42/clients/detectionlists.py
@@ -1,3 +1,23 @@
+from py42.sdk.queries.query_filter import filter_attributes
+
+
+class RiskTags(object):
+    """The available RiskTags for :meth:`~py42.modules.detectionlists.DetectionListsModule.add_user_risk_tags()`
+    and :meth:`~py42.modules.detectionlists.DetectionListsModule.remove_user_risk_tags()`"""
+
+    FLIGHT_RISK = u"FLIGHT_RISK"
+    HIGH_IMPACT_EMPLOYEE = u"HIGH_IMPACT_EMPLOYEE"
+    ELEVATED_ACCESS_PRIVILEGES = u"ELEVATED_ACCESS_PRIVILEGES"
+    PERFORMANCE_CONCERNS = u"PERFORMANCE_CONCERNS"
+    SUSPICIOUS_SYSTEM_ACTIVITY = u"SUSPICIOUS_SYSTEM_ACTIVITY"
+    POOR_SECURITY_PRACTICES = u"POOR_SECURITY_PRACTICES"
+    CONTRACT_EMPLOYEE = u"CONTRACT_EMPLOYEE"
+
+    @staticmethod
+    def choices():
+        return filter_attributes(RiskTags)
+
+
 class DetectionListsModule(object):
     def __init__(
         self,