+++ tests/modules/test_detectionlists.py
@@ -18,6 +18,21 @@ def mock_detection_list_user_client(mocker):
 TEST_USER_ID = "12345"
 
 
+class TestRiskTags(object):
+    def test_choices_returns_expected_set(self):
+        choices = RiskTags.choices()
+        valid_set = {
+            "FLIGHT_RISK",
+            "HIGH_IMPACT_EMPLOYEE",
+            "ELEVATED_ACCESS_PRIVILEGES",
+            "PERFORMANCE_CONCERNS",
+            "SUSPICIOUS_SYSTEM_ACTIVITY",
+            "POOR_SECURITY_PRACTICES",
+            "CONTRACT_EMPLOYEE",
+        }
+        assert set(choices) == valid_set
+
+
 class TestDetectionListModule(object):
     def test_departing_employee_call_get_departing_employee_client_with_expected_values(
         self, mock_microservice_client_factory